refactor(service): reorganize the backup module
- New submodules `config` and `service`.
- Remove "backup" as a default feature.
- `backup::config` module is not feature-gated.
