+++ crates/walrus-service/src/lib.rs
@@ -12,7 +12,6 @@ pub mod node;
 #[cfg(feature = "deploy")]
 pub mod testbed;
 
-#[cfg(feature = "backup")]
 pub mod backup;
 
 #[cfg(any(feature = "client", feature = "node"))]
